+++ tests/parse/test_sccmec.py
@@ -17,11 +17,11 @@ def test_parse_sccmec_results(saureus_sccmec_path):
             "type": "IV",
             "subtype": "multiple",
             "mecA": "+",
-            "targets": "ccrA2,ccrB2,IS431,IS431_1,IS431_2,IS1272,mecA,mecR1",
-            "regions": "IVa,IVn",
-            "coverage": "96.31,83.93",
-            "hits": "27,25",
-            "target_comment": "",
+            "targets": ["ccrA2", "ccrB2", "IS431", "IS431_1", "IS431_2", "IS1272", "mecA", "mecR1"],
+            "regions": ["IVa", "IVn"],
+            "coverage": [96.31, 83.93],
+            "hits": [27,25],
+            "target_comment": None,
             "region_comment": "Found matches for multiple types including: IVa, IVn",
             "comment": "The type was determined based on matches to multiple subtypes of the same type"
         }